pyproject.toml
[project]
name = "murasame-bot"
version = "0.1.0"
description = "murasame-bot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[httpx]>=2.4.4",
    "nonebot2[websockets]>=2.4.4",
    "nonebot2[fastapi]",
    "nonebot-adapter-qq>=1.6.6",
    "nonebot-adapter-onebot>=2.4.6",
    "discord.py"
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo", "single_session"]

[tool.nonebot.adapters]
#nonebot-adapter-qq = [
#    { name = "QQ", module_name = "nonebot.adapters.qq" }
#]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
